requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
install==1.3.5
pycodestyle==2.9.1
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
whitenoise==6.2.0
